# src/main/resources/swagger.yaml
swagger: '2.0'
info:
  description: DataStation API
  version: v0.01
  title: DataStation API
  termsOfService: Terms of service
  contact:
    name: N.A
    url: N.A
    email: N.A
  license:
    name: License of API
    url: API license URL
host: '192.168.210.131:7578'
basePath: /
...
paths:
...
  /weather/download:
    get:
      tags:
        - weather-controller
      summary: downloadMultipleWeather
      description: 날씨 데이터셋 다운로드; dataset 정보는 192.168.210.142 dataset_info 참고
      operationId: downloadMultipleWeatherUsingGET
      produces:
        - application/json
      parameters:
        - name: dataset
          in: query
          description: w-avg-ta, w-max-ta, w-min-ta, w-prob-rn, w-sum-rn, w-dd-mefs, w-max-wd, w-avg-ws, w-avg-rhm, w-avg-tca, a-pm10, a-pm25, a-o3, s-01-swind, s-02-hrain, s-03-cold, s-04-dry, s-05-ssurge, s-06-hsea, s-07-typoon, s-08-hsnow, s-09-ydust, s-10-resv
          required: true
          type: array
          items:
            type: string
          collectionFormat: multi
        - name: format
          in: query
          description: format
          required: false
          type: string
          default: csv
        - name: from
          in: query
          description: \"yyyyMMdd\" format String
          required: false
          type: string
          x-example: '20170101'
        - name: to
          in: query
          description: \"yyyyMMdd\" format String
          required: false
          type: string
          x-example: '20191231'
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
        - JWT:
            - global
      deprecated: false
  /weather/spss/download:
    get:
      tags:
        - weather-controller
      summary: downloadFullWeatherSPSS
      description: 날씨 데이터셋 CSV 다운로드(SPSS); dataset 정보는 192.168.210.142 dataset_info 참고
      operationId: downloadFullWeatherSPSSUsingGET
      produces:
        - '*/*'
      parameters:
        - name: areacode
          in: query
          description: areacode; dd >= 10 && dd <= 90; 192.168.210.142 weather_area_info 참고
          required: false
          type: array
          items:
            type: integer
            format: int32
          collectionFormat: multi
        - name: from
          in: query
          description: \"yyyyMMdd\" format String
          required: false
          type: string
          x-example: '20170101'
        - name: to
          in: query
          description: \"yyyyMMdd\" format String
          required: false
          type: string
          x-example: '20181231'
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
        - JWT:
            - global
      deprecated: false
  '/weather/{dataset}/download':
    get:
      tags:
        - weather-controller
      summary: downloadWeather
      description: 날씨 데이터셋 다운로드; dataset 정보는 192.168.210.142 dataset_info 참고
      operationId: downloadWeatherUsingGET
      produces:
        - application/json
      parameters:
        - name: areacode
          in: query
          description: areacode; dd >= 10 && dd <= 90; 192.168.210.142 weather_area_info 참고
          required: false
          type: array
          items:
            type: integer
            format: int32
          collectionFormat: multi
        - name: dataset
          in: path
          description: w-avg-ta, w-max-ta, w-min-ta, w-prob-rn, w-sum-rn, w-dd-mefs, w-max-wd, w-avg-ws, w-avg-rhm, w-avg-tca, a-pm10, a-pm25, a-o3, s-01-swind, s-02-hrain, s-03-cold, s-04-dry, s-05-ssurge, s-06-hsea, s-07-typoon, s-08-hsnow, s-09-ydust, s-10-resv
          required: true
          type: string
        - name: format
          in: query
          description: format
          required: false
          type: string
          default: csv
        - name: from
          in: query
          description: \"yyyyMMdd\" format String
          required: false
          type: string
          x-example: '20170101'
        - name: to
          in: query
          description: \"yyyyMMdd\" format String
          required: false
          type: string
          x-example: '20181231'
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
        - JWT:
            - global
      deprecated: false
securityDefinitions:
  JWT:
    type: apiKey
    name: Authorization
    in: header
...